Add rescue for RecordNotFound errors

+++ contact_list.rb
@@ -28,16 +28,25 @@ class ContactList
     puts "Please give me a valid email:"
     email = STDIN.gets.chomp
     created = Contact.create(name: name, email: email)
-    puts created
+    if created.errors[:email].any?
+      puts "Could not create contact: email address already exists"
+    else
+      puts "Contact created!"
+    end
   elsif ARGV[0] == "show"
     puts "Please give me a contact ID:"
     id = STDIN.gets.to_i
-    contact = Contact.find(id)
-    if contact.nil? || id < 1
-      puts "Contact with id: #{id} cannot be found"
-    else
+    begin
+      contact = Contact.find(id)
+    rescue ActiveRecord::RecordNotFound
+      contact = false
+    end
+    if contact
+      puts "Here is your contact:"
       puts contact.name
       puts contact.email
+    else
+      puts "Contact with id: #{id} cannot be found"
     end
   elsif ARGV[0] == "search"
     puts "Please enter a search term:"
@@ -55,19 +64,39 @@ class ContactList
     new_name = STDIN.gets.chomp
     puts "Please enter a new email:"
     new_email = STDIN.gets.chomp
-    contact = Contact.find(id)
-    contact.name = new_name
-    contact.email = new_email
-    contact.save
+    begin
+      contact = Contact.find(id)
+    rescue ActiveRecord::RecordNotFound
+      contact = false
+    end
+    if contact
+      contact.name = new_name
+      contact.email = new_email
+      contact.save
+      puts "Contact added"
+    else
+      puts "Contact with id: #{id} cannot be found"
+    end
   elsif ARGV[0] == "destroy"
     puts "Please enter a contact ID:"
     id = STDIN.gets.to_i
-    puts "Are you sure?"
-    answer = STDIN.gets.chomp
-    answer.downcase!
-    if answer == "y" || answer == "yes"
+    begin
       contact = Contact.find(id)
-      contact.destroy
+    rescue ActiveRecord::RecordNotFound
+      contact = false
+    end
+    if contact
+      puts "Are you sure?"
+      answer = STDIN.gets.chomp
+      answer.downcase!
+      if answer == "y" || answer == "yes"
+        contact.destroy
+        puts "Contact was deleted :)"
+      else
+        puts "Contact was not deleted"
+      end
+    else
+      puts "Contact with id: #{id} cannot be found"
     end
   else
     puts "idk"
